+++ src/lib/index.js
@@ -36,7 +36,6 @@ export { default as ContactForm } from '$lib/molecules/contact-form.svelte';
 // =========== ORGANISMS ===========
 
 export { default as Menu } from '$lib/organisms/menu.svelte';   
-export { default as Navigation } from '$lib/organisms/navigation.svelte';
 export { default as SlideCards } from '$lib/organisms/slide-cards.svelte';
 export { default as Footer } from '$lib/organisms/footer.svelte';
 export { default as Agenda } from '$lib/organisms/agenda.svelte';
